requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.4.0
boto3==1.42.17
botocore==1.42.17
brotli==1.2.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
cryptography==46.0.5
distro==1.9.0
Django==4.2.30
django-environ==0.13.0
django-jazzmin==3.0.1
djangorestframework==3.16.1
exceptiongroup==1.3.1
frozenlist==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-genai==1.47.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jiter==0.13.0
jmespath==1.0.1
lxml==6.0.2
multidict==6.7.1
nest-asyncio2==1.7.2
openai==2.24.0
pillow==10.2.0
primp==0.15.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-telegram-bot==22.5
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
sqlparse==0.5.3
tenacity==9.1.2
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==1.26.20
websockets==15.0.1
yarl==1.22.0
